Prep work for modifiers + bug fixes
Tool and bookmark item buttons are now grouped and can be paginated. The bookmarked items have also now moved to below the tools to create space for modifiers. Fixed bug with tooltips not allowing items to render below them. JEI has been added to the build.gradle to allow for easier testing.