A lot of modifier work
Added a "back" button in the modifier screen to go back to the list. The - and + buttons are now arrows and a preview of the tool is shown for each button. How modifiers are stored were completely rewritten to better emulate an actual tool and for more fine control and verification. A few other visual changes were also made.